python-scripts/requirements.txt
mysql-connector-python
pymongo
redis
flask